requirements.txt
anyio==3.6.1
asgiref==3.5.1
click==8.1.3
fastapi==0.78.0
h11==0.13.0
idna==3.3
ImageHash==4.2.1
loguru==0.6.0
numpy==1.22.3
Pillow==9.1.0
pydantic==1.9.0
python-multipart==0.0.5
PyWavelets==1.3.0
scipy==1.8.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing-extensions==4.2.0
uvicorn==0.17.6
